tests/wpt/web-platform-tests/dom/nodes/Document-contentType/contentType/createDocument.html

<!DOCTYPE html>
<title>document.implementation.createDocument: document.contentType === 'application/xml'</title>
<script src="/resources/testharness.js"></script>
<script src="/resources/testharnessreport.js"></script>
<div id=log></div>
<script>
test(function() {
  var doc = document.implementation.createDocument("http://www.w3.org/1999/xhtml", "html", null);
  assert_equals(doc.contentType, "application/xml");
});
</script>
