LayoutTests/imported/w3c/web-platform-tests/dom/nodes/Document-contentType/contentType/createDocument-expected.txt


FAIL document.implementation.createDocument: document.contentType === 'application/xml' assert_equals: expected "application/xml" but got "application/xhtml+xml"
